HT_20/sears/requirements/development.txt
-r base.txt

django-extensions==3.2.3
django-debug-toolbar==4.2.0
flake8==6.1.0
ipython==8.19.0
isort==5.13.2